# backend/tests/openapi.snapshot.json
{
 "components": {
  "schemas": {
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "title": "Detail",
      "type": "array"
     }
    },
    "title": "HTTPValidationError",
    "type": "object"
   },
...
   "PreflightRequest": {
    "properties": {
     "selection": {
      "$ref": "#/components/schemas/TaskSelection"
     },
     "settings": {
      "additionalProperties": true,
      "title": "Settings",
      "type": "object"
     }
    },
    "title": "PreflightRequest",
    "type": "object"
   },
   "PreflightResult": {
    "description": "TSK-04: items that aren't ready are skipped, never failed.",
    "properties": {
     "derived_root_required": {
      "default": false,
      "title": "Derived Root Required",
      "type": "boolean"
     },
     "missing": {
      "additionalProperties": {
       "type": "integer"
      },
      "title": "Missing",
      "type": "object"
     },
     "n_ready": {
      "title": "N Ready",
      "type": "integer"
     },
     "n_selected": {
      "title": "N Selected",
      "type": "integer"
     },
     "ready_item_ids": {
      "items": {
       "type": "string"
      },
      "title": "Ready Item Ids",
      "type": "array"
     },
     "suggestions": {
      "items": {
       "$ref": "#/components/schemas/Suggestion"
      },
      "title": "Suggestions",
      "type": "array"
     }
    },
    "required": [
     "n_selected",
     "n_ready",
     "missing",
     "suggestions"
    ],
    "title": "PreflightResult",
    "type": "object"
   },
...
   "SegProducer": {
    "description": "The task run that produced a segmentation set (ADR-0015).",
    "properties": {
     "run_id": {
      "title": "Run Id",
      "type": "string"
     },
     "settings_hash": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Settings Hash"
     },
     "task_id": {
      "title": "Task Id",
      "type": "string"
     },
     "version": {
      "title": "Version",
      "type": "string"
     }
    },
    "required": [
     "task_id",
     "version",
     "run_id"
    ],
    "title": "SegProducer",
    "type": "object"
   },
...
   "SegmentationInfo": {
    "properties": {
     "created_at": {
      "default": "",
      "title": "Created At",
      "type": "string"
     },
     "is_default": {
      "default": false,
      "title": "Is Default",
      "type": "boolean"
     },
     "kind": {
      "default": "imported",
      "enum": [
       "imported",
       "task",
       "manual"
      ],
      "title": "Kind",
      "type": "string"
     },
     "label_mapping": {
      "additionalProperties": {
       "type": "integer"
      },
      "title": "Label Mapping",
      "type": "object"
     },
     "n_items": {
      "default": 0,
      "title": "N Items",
      "type": "integer"
     },
     "name": {
      "default": "",
      "title": "Name",
      "type": "string"
     },
     "producer": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/SegProducer"
       },
       {
        "type": "null"
       }
      ]
     },
     "seg_id": {
      "pattern": "^[a-z0-9][a-z0-9_-]{0,63}$",
      "title": "Seg Id",
      "type": "string"
     },
     "unmatched": {
      "items": {
       "type": "integer"
      },
      "title": "Unmatched",
      "type": "array"
     }
    },
    "required": [
     "seg_id"
    ],
    "title": "SegmentationInfo",
    "type": "object"
   },
   "SegmentationPatch": {
    "properties": {
     "label_mapping": {
      "anyOf": [
       {
        "additionalProperties": {
         "type": "integer"
        },
        "type": "object"
       },
       {
        "type": "null"
       }
      ],
      "title": "Label Mapping"
     },
     "name": {
      "anyOf": [
       {
        "maxLength": 200,
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Name"
     }
    },
    "title": "SegmentationPatch",
    "type": "object"
   },
...
   "Suggestion": {
    "properties": {
     "reason": {
      "title": "Reason",
      "type": "string"
     },
     "task_id": {
      "title": "Task Id",
      "type": "string"
     }
    },
    "required": [
     "task_id",
     "reason"
    ],
    "title": "Suggestion",
    "type": "object"
   },
...
   "TaskEstimate": {
    "description": "TSK-05.",
    "properties": {
     "basis": {
      "enum": [
       "sample",
       "manifest",
       "unknown"
      ],
      "title": "Basis",
      "type": "string"
     },
     "estimated_total_s": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Estimated Total S"
     },
     "n_skipped": {
      "title": "N Skipped",
      "type": "integer"
     },
     "n_units": {
      "title": "N Units",
      "type": "integer"
     },
     "output_bytes": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Output Bytes"
     },
     "sample_errors": {
      "items": {
       "type": "string"
      },
      "title": "Sample Errors",
      "type": "array"
     },
     "sample_item_ids": {
      "items": {
       "type": "string"
      },
      "title": "Sample Item Ids",
      "type": "array"
     },
     "seconds_per_item": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Seconds Per Item"
     }
    },
    "required": [
     "n_units",
     "n_skipped",
     "seconds_per_item",
     "estimated_total_s",
     "basis"
    ],
    "title": "TaskEstimate",
    "type": "object"
   },
   "TaskInfo": {
    "description": "API-42 row: a loaded manifest plus where it came from and whether it can run now.",
    "properties": {
     "available": {
      "default": true,
      "title": "Available",
      "type": "boolean"
     },
     "manifest": {
      "$ref": "#/components/schemas/TaskManifest"
     },
     "manifest_hash": {
      "title": "Manifest Hash",
      "type": "string"
     },
     "runner_online": {
      "anyOf": [
       {
        "type": "boolean"
       },
       {
        "type": "null"
       }
      ],
      "title": "Runner Online"
     },
     "settings_schema_url": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Settings Schema Url"
     },
     "source": {
      "enum": [
       "builtin",
       "plugins_root"
      ],
      "title": "Source",
      "type": "string"
     },
     "unavailable_reason": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Unavailable Reason"
     }
    },
    "required": [
     "manifest",
     "source",
     "manifest_hash"
    ],
    "title": "TaskInfo",
    "type": "object"
   },
   "TaskItemError": {
    "properties": {
     "item_id": {
      "title": "Item Id",
      "type": "string"
     },
     "message": {
      "default": "",
      "title": "Message",
      "type": "string"
     },
     "status": {
      "enum": [
       "ok",
       "failed",
       "skipped"
      ],
      "title": "Status",
      "type": "string"
     }
    },
    "required": [
     "item_id",
     "status"
    ],
    "title": "TaskItemError",
    "type": "object"
   },
...
   "TaskRunDetail": {
    "properties": {
     "attempts": {
      "default": 0,
      "title": "Attempts",
      "type": "integer"
     },
     "counts": {
      "$ref": "#/components/schemas/TaskRunCounts"
     },
     "created_at": {
      "title": "Created At",
      "type": "string"
     },
     "detail_url": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Detail Url"
     },
     "error": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Error"
     },
     "finished_at": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Finished At"
     },
     "inputs": {
      "items": {
       "$ref": "#/components/schemas/TaskRunInput"
      },
      "title": "Inputs",
      "type": "array"
     },
     "item_ids": {
      "items": {
       "type": "string"
      },
      "title": "Item Ids",
      "type": "array"
     },
     "job_id": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Job Id"
     },
     "name": {
      "title": "Name",
      "type": "string"
     },
     "output_dir": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Output Dir"
     },
     "outputs": {
      "items": {
       "$ref": "#/components/schemas/TaskRunOutput"
      },
      "title": "Outputs",
      "type": "array"
     },
     "progress": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/TaskRunProgress"
       },
       {
        "type": "null"
       }
      ]
     },
     "reviewer": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Reviewer"
     },
     "run_id": {
      "title": "Run Id",
      "type": "string"
     },
     "runtime": {
      "enum": [
       "builtin",
       "external"
      ],
      "title": "Runtime",
      "type": "string"
     },
     "selection": {
      "$ref": "#/components/schemas/TaskSelection"
     },
     "settings": {
      "additionalProperties": true,
      "title": "Settings",
      "type": "object"
     },
     "settings_hash": {
      "title": "Settings Hash",
      "type": "string"
     },
     "started_at": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Started At"
     },
     "status": {
      "enum": [
       "queued",
       "waiting_for_runner",
       "running",
       "completed",
       "completed_with_errors",
       "failed",
       "cancelled",
       "interrupted"
      ],
      "title": "Status",
      "type": "string"
     },
     "task": {
      "$ref": "#/components/schemas/TaskRef"
     },
     "versions": {
      "additionalProperties": {
       "type": "string"
      },
      "title": "Versions",
      "type": "object"
     }
    },
    "required": [
     "run_id",
     "task",
     "name",
     "status",
     "created_at",
     "runtime",
     "settings",
     "settings_hash",
     "selection"
    ],
    "title": "TaskRunDetail",
    "type": "object"
   },
...
   "TaskRunOutput": {
    "description": "A registered output (TSK-09).",
    "properties": {
     "detail": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Detail"
     },
     "item_id": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Item Id"
     },
     "kind": {
      "title": "Kind",
      "type": "string"
     },
     "ref": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Ref"
     },
     "seg_id": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Seg Id"
     },
     "sha256": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Sha256"
     }
    },
    "required": [
     "kind"
    ],
    "title": "TaskRunOutput",
    "type": "object"
   },
...
   "TaskRunRequest": {
    "properties": {
     "name": {
      "anyOf": [
       {
        "maxLength": 200,
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Name"
     },
     "selection": {
      "$ref": "#/components/schemas/TaskSelection"
     },
     "settings": {
      "additionalProperties": true,
      "title": "Settings",
      "type": "object"
     },
     "task_id": {
      "title": "Task Id",
      "type": "string"
     }
    },
    "required": [
     "task_id"
    ],
    "title": "TaskRunRequest",
    "type": "object"
   },
   "TaskRunStarted": {
    "properties": {
     "job_id": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Job Id"
     },
     "run_id": {
      "title": "Run Id",
      "type": "string"
     },
     "status": {
      "enum": [
       "queued",
       "waiting_for_runner",
       "running",
       "completed",
       "completed_with_errors",
       "failed",
       "cancelled",
       "interrupted"
      ],
      "title": "Status",
      "type": "string"
     }
    },
    "required": [
     "run_id",
     "job_id",
     "status"
    ],
    "title": "TaskRunStarted",
    "type": "object"
   },
   "TaskRunSummary": {
    "properties": {
     "counts": {
      "$ref": "#/components/schemas/TaskRunCounts"
     },
     "created_at": {
      "title": "Created At",
      "type": "string"
     },
     "error": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Error"
     },
     "finished_at": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Finished At"
     },
     "job_id": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Job Id"
     },
     "name": {
      "title": "Name",
      "type": "string"
     },
     "reviewer": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Reviewer"
     },
     "run_id": {
      "title": "Run Id",
      "type": "string"
     },
     "started_at": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Started At"
     },
     "status": {
      "enum": [
       "queued",
       "waiting_for_runner",
       "running",
       "completed",
       "completed_with_errors",
       "failed",
       "cancelled",
       "interrupted"
      ],
      "title": "Status",
      "type": "string"
     },
     "task": {
      "$ref": "#/components/schemas/TaskRef"
     }
    },
    "required": [
     "run_id",
     "task",
     "name",
     "status",
     "created_at"
    ],
    "title": "TaskRunSummary",
    "type": "object"
   },
...
   "TaskValidateRequest": {
    "properties": {
     "settings": {
      "additionalProperties": true,
      "title": "Settings",
      "type": "object"
     }
    },
    "title": "TaskValidateRequest",
    "type": "object"
   },
   "TaskValidateResult": {
    "properties": {
     "issues": {
      "items": {
       "$ref": "#/components/schemas/SettingsIssue"
      },
      "title": "Issues",
      "type": "array"
     },
     "ok": {
      "title": "Ok",
      "type": "boolean"
     },
     "settings": {
      "anyOf": [
       {
        "additionalProperties": true,
        "type": "object"
       },
       {
        "type": "null"
       }
      ],
      "title": "Settings"
     },
     "settings_hash": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Settings Hash"
     }
    },
    "required": [
     "ok",
     "issues"
    ],
    "title": "TaskValidateResult",
    "type": "object"
   },
...
  }
 },
 "info": {
  "title": "Radiology Workbench",
  "version": "3.0.0.dev0"
 },
 "openapi": "3.1.0",
 "paths": {
...
  "/api/v1/projects/{pid}/segmentations": {
   "get": {
    "operationId": "list_segmentations_api_v1_projects__pid__segmentations_get",
    "parameters": [
     {
      "in": "path",
      "name": "pid",
      "required": true,
      "schema": {
       "title": "Pid",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "items": {
          "$ref": "#/components/schemas/SegmentationInfo"
         },
         "title": "Response List Segmentations Api V1 Projects  Pid  Segmentations Get",
         "type": "array"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "List Segmentations",
    "tags": [
     "segmentations"
    ]
   }
  },
  "/api/v1/projects/{pid}/segmentations/{seg}": {
   "patch": {
    "operationId": "patch_segmentation_api_v1_projects__pid__segmentations__seg__patch",
    "parameters": [
     {
      "in": "path",
      "name": "pid",
      "required": true,
      "schema": {
       "title": "Pid",
       "type": "string"
      }
     },
     {
      "in": "path",
      "name": "seg",
      "required": true,
      "schema": {
       "title": "Seg",
       "type": "string"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/SegmentationPatch"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/SegmentationInfo"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Patch Segmentation",
    "tags": [
     "segmentations"
    ]
   }
  },
  "/api/v1/projects/{pid}/task-runs": {
   "get": {
    "operationId": "list_runs_api_v1_projects__pid__task_runs_get",
    "parameters": [
     {
      "in": "path",
      "name": "pid",
      "required": true,
      "schema": {
       "title": "Pid",
       "type": "string"
      }
     },
     {
      "in": "query",
      "name": "task",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Task"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "items": {
          "$ref": "#/components/schemas/TaskRunSummary"
         },
         "title": "Response List Runs Api V1 Projects  Pid  Task Runs Get",
         "type": "array"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "List Runs",
    "tags": [
     "tasks"
    ]
   },
   "post": {
    "operationId": "start_run_api_v1_projects__pid__task_runs_post",
    "parameters": [
     {
      "in": "path",
      "name": "pid",
      "required": true,
      "schema": {
       "title": "Pid",
       "type": "string"
      }
     },
     {
      "in": "header",
      "name": "X-Reviewer",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Reviewer"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/TaskRunRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "202": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/TaskRunStarted"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Start Run",
    "tags": [
     "tasks"
    ]
   }
  },
  "/api/v1/projects/{pid}/task-runs/{rid}": {
   "get": {
    "operationId": "get_run_api_v1_projects__pid__task_runs__rid__get",
    "parameters": [
     {
      "in": "path",
      "name": "pid",
      "required": true,
      "schema": {
       "title": "Pid",
       "type": "string"
      }
     },
     {
      "in": "path",
      "name": "rid",
      "required": true,
      "schema": {
       "title": "Rid",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/TaskRunDetail"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Get Run",
    "tags": [
     "tasks"
    ]
   }
  },
  "/api/v1/projects/{pid}/task-runs/{rid}/cancel": {
   "post": {
    "operationId": "cancel_run_api_v1_projects__pid__task_runs__rid__cancel_post",
    "parameters": [
     {
      "in": "path",
      "name": "pid",
      "required": true,
      "schema": {
       "title": "Pid",
       "type": "string"
      }
     },
     {
      "in": "path",
      "name": "rid",
      "required": true,
      "schema": {
       "title": "Rid",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/TaskRunDetail"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Cancel Run",
    "tags": [
     "tasks"
    ]
   }
  },
  "/api/v1/projects/{pid}/task-runs/{rid}/errors": {
   "get": {
    "operationId": "run_errors_api_v1_projects__pid__task_runs__rid__errors_get",
    "parameters": [
     {
      "in": "path",
      "name": "pid",
      "required": true,
      "schema": {
       "title": "Pid",
       "type": "string"
      }
     },
     {
      "in": "path",
      "name": "rid",
      "required": true,
      "schema": {
       "title": "Rid",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "items": {
          "$ref": "#/components/schemas/TaskItemError"
         },
         "title": "Response Run Errors Api V1 Projects  Pid  Task Runs  Rid  Errors Get",
         "type": "array"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Run Errors",
    "tags": [
     "tasks"
    ]
   }
  },
  "/api/v1/projects/{pid}/task-runs/{rid}/outputs": {
   "get": {
    "operationId": "run_outputs_api_v1_projects__pid__task_runs__rid__outputs_get",
    "parameters": [
     {
      "in": "path",
      "name": "pid",
      "required": true,
      "schema": {
       "title": "Pid",
       "type": "string"
      }
     },
     {
      "in": "path",
      "name": "rid",
      "required": true,
      "schema": {
       "title": "Rid",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "items": {
          "$ref": "#/components/schemas/TaskRunOutput"
         },
         "title": "Response Run Outputs Api V1 Projects  Pid  Task Runs  Rid  Outputs Get",
         "type": "array"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Run Outputs",
    "tags": [
     "tasks"
    ]
   }
  },
  "/api/v1/projects/{pid}/task-runs/{rid}/resume": {
   "post": {
    "operationId": "resume_run_api_v1_projects__pid__task_runs__rid__resume_post",
    "parameters": [
     {
      "in": "path",
      "name": "pid",
      "required": true,
      "schema": {
       "title": "Pid",
       "type": "string"
      }
     },
     {
      "in": "path",
      "name": "rid",
      "required": true,
      "schema": {
       "title": "Rid",
       "type": "string"
      }
     }
    ],
    "responses": {
     "202": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/TaskRunStarted"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Resume Run",
    "tags": [
     "tasks"
    ]
   }
  },
  "/api/v1/projects/{pid}/tasks/{tid}/estimate": {
   "post": {
    "operationId": "estimate_api_v1_projects__pid__tasks__tid__estimate_post",
    "parameters": [
     {
      "in": "path",
      "name": "pid",
      "required": true,
      "schema": {
       "title": "Pid",
       "type": "string"
      }
     },
     {
      "in": "path",
      "name": "tid",
      "required": true,
      "schema": {
       "title": "Tid",
       "type": "string"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/PreflightRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/TaskEstimate"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Estimate",
    "tags": [
     "tasks"
    ]
   }
  },
  "/api/v1/projects/{pid}/tasks/{tid}/preflight": {
   "post": {
    "operationId": "preflight_api_v1_projects__pid__tasks__tid__preflight_post",
    "parameters": [
     {
      "in": "path",
      "name": "pid",
      "required": true,
      "schema": {
       "title": "Pid",
       "type": "string"
      }
     },
     {
      "in": "path",
      "name": "tid",
      "required": true,
      "schema": {
       "title": "Tid",
       "type": "string"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/PreflightRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/PreflightResult"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Preflight",
    "tags": [
     "tasks"
    ]
   }
  },
...
  "/api/v1/tasks/{tid}": {
   "get": {
    "operationId": "get_task_api_v1_tasks__tid__get",
    "parameters": [
     {
      "in": "path",
      "name": "tid",
      "required": true,
      "schema": {
       "title": "Tid",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/TaskInfo"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Get Task",
    "tags": [
     "tasks"
    ]
   }
  },
  "/api/v1/tasks/{tid}/validate": {
   "post": {
    "operationId": "validate_task_api_v1_tasks__tid__validate_post",
    "parameters": [
     {
      "in": "path",
      "name": "tid",
      "required": true,
      "schema": {
       "title": "Tid",
       "type": "string"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/TaskValidateRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/TaskValidateResult"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Validate Task",
    "tags": [
     "tasks"
    ]
   }
  }
 }
}